requirements.txt
keyboard>=0.13.5
pystray>=0.19.5
Pillow>=10.0.0
pywin32>=306
customtkinter>=5.2.0
